third_party/blink/web_tests/external/wpt/css/css-typed-om/historical-expected.txt

This is a testharness.js-based test.
[FAIL] Support for CSSPositionValue
  assert_false: expected false got true
[FAIL] toRGB should not exist on CSSColorValue
  assert_false: expected false got true
[FAIL] toHSL should not exist on CSSColorValue
  assert_false: expected false got true
[FAIL] toHWB should not exist on CSSColorValue
  assert_false: expected false got true
Harness: the test ran to completion.
